FIX Split NGSI Service into NGSI and Device registration
