+++ lib/fiware-iotagent-lib.js
@@ -26,6 +26,7 @@
 var async = require('async'),
     apply = async.apply,
     ngsi = require('./services/ngsiService'),
+    deviceService = require('./services/deviceService'),
     groupConfig = require('./services/groupService'),
     security = require('./services/securityService'),
     contextServer = require('./services/northboundServer'),
@@ -73,6 +74,7 @@ function doActivate(newConfig, callback) {
     async.series([
         async.apply(registry.init, newConfig),
         async.apply(ngsi.init, registry, groupRegistry, newConfig),
+        async.apply(deviceService.init, registry, groupRegistry, newConfig),
         async.apply(groupConfig.init, groupRegistry, newConfig),
         async.apply(security.init, newConfig),
         async.apply(contextServer.start, newConfig)
@@ -112,15 +114,15 @@ function deactivate(callback) {
 
 exports.activate = activate;
 exports.deactivate = deactivate;
-exports.register = ngsi.register;
-exports.updateRegister = ngsi.updateRegister;
-exports.unregister = ngsi.unregister;
+exports.register = deviceService.register;
+exports.updateRegister = deviceService.updateRegister;
+exports.unregister = deviceService.unregister;
 exports.query = ngsi.query;
 exports.update = ngsi.update;
 exports.setCommandResult = ngsi.setCommandResult;
-exports.listDevices = ngsi.listDevices;
-exports.getDevice = ngsi.getDevice;
-exports.getDeviceByName = ngsi.getDeviceByName;
+exports.listDevices = deviceService.listDevices;
+exports.getDevice = deviceService.getDevice;
+exports.getDeviceByName = deviceService.getDeviceByName;
 exports.getConfiguration = groupConfig.get;
 exports.setDataUpdateHandler = contextServer.setUpdateHandler;
 exports.setCommandHandler = contextServer.setCommandHandler;
