+++ lib/services/contextServer.js
@@ -27,6 +27,7 @@ var async = require('async'),
     logger = require('fiware-node-logger'),
     errors = require('../errors'),
     ngsi = require('./ngsiService'),
+    deviceService = require('./deviceService'),
     revalidator = require('revalidator'),
     _ = require('underscore'),
     context = {
@@ -271,7 +272,7 @@ function generateUpdateActions(req, contextElement, callback) {
     }
 
     async.waterfall([
-        apply(ngsi.getDeviceByName, contextElement.id),
+        apply(deviceService.getDeviceByName, contextElement.id),
         splitUpdates,
         createActionsArray
     ], callback);
@@ -358,7 +359,7 @@ function handleQuery(req, res, next) {
     }
 
     function createQueryRequests(attributes, contextEntity, callback) {
-        ngsi.getDeviceByName(contextEntity.id, function handleFindDevice(error, device) {
+        deviceService.getDeviceByName(contextEntity.id, function handleFindDevice(error, device) {
             var executeCompleteAttributes = apply(
                     completeAttributes,
                     attributes,
