+++ lib/services/deviceProvisioningServer.js
@@ -25,6 +25,7 @@
 var async = require('async'),
     restUtils = require('./restUtils'),
     ngsi = require('./ngsiService'),
+    deviceService = require('./deviceService'),
     logger = require('fiware-node-logger'),
     errors = require('../errors'),
     _ = require('underscore'),
@@ -67,7 +68,7 @@ function handleProvision(req, res, next) {
 
     function registerDevice(service, subservice, body, callback) {
         /*jshint sub:true */
-        ngsi.register({
+        deviceService.register({
                 id: body['name'],
                 type: body['entity_type'],
                 name: body['entity_name'],
@@ -97,7 +98,7 @@ function handleProvision(req, res, next) {
  * Express middleware that retrieves the complete set of provisioned devices (in JSON format).
  */
 function handleListDevices(req, res, next) {
-    ngsi.listDevices(
+    deviceService.listDevices(
         req.headers['fiware-service'],
         req.headers['fiware-servicepath'],
         req.query.limit,
@@ -138,7 +139,7 @@ function toProvisioningAPIFormat(device) {
  * JSON format.
  */
 function handleGetDevice(req, res, next) {
-    ngsi.getDevice(req.params.deviceId, function(error, device) {
+    deviceService.getDevice(req.params.deviceId, function(error, device) {
         if (error) {
             next(error);
         } else if (device) {
@@ -153,7 +154,7 @@ function handleGetDevice(req, res, next) {
  * This middleware handles the removal of a particular device specified with the deviceId.
  */
 function handleRemoveDevice(req, res, next) {
-    ngsi.unregister(req.params.deviceId, function(error) {
+    deviceService.unregister(req.params.deviceId, function(error) {
         if (error) {
             next(error);
         } else {
@@ -169,7 +170,7 @@ function handleUpdateDevice(req, res, next) {
     if (req.body.name) {
         next(new errors.BadRequest('Can\'t change the ID of a preprovisioned device'));
     } else {
-        ngsi.getDevice(req.params.deviceId, function(error, device) {
+        deviceService.getDevice(req.params.deviceId, function(error, device) {
             if (error) {
                 next(error);
             } else if (device) {
@@ -179,7 +180,7 @@ function handleUpdateDevice(req, res, next) {
                     device[provisioningAPITranslation[pairs[i][0]]] = pairs[i][1];
                 }
 
-                ngsi.updateRegister(device, function handleDeviceUpdate(error) {
+                deviceService.updateRegister(device, function handleDeviceUpdate(error) {
                     if (error) {
                         next(error);
                     } else {
