+++ lib/services/ngsiParser.js
@@ -201,5 +201,26 @@ function readBody(parseFunction) {
     };
 }
 
+/**
+ * Given a NGSI Body, determines whether it contains any NGSI error.
+ *
+ * @param {String} body             String representing a NGSI body in JSON format.
+ * @return {Number|*}
+ */
+function getErrorField(body) {
+    var errorField = body.errorCode ||
+        body.orionError;
+
+    if (body && body.contextResponses &&
+        body.contextResponses[0] && body.contextResponses[0] &&
+        body.contextResponses[0].statusCode && body.contextResponses[0].statusCode.code !== '200') {
+        errorField = body.contextResponses[0].statusCode.reasonPhrase +
+        ': ' + body.contextResponses[0].statusCode.details;
+    }
+
+    return errorField;
+}
+
 exports.readUpdateBody = readBody(readUpdatePayload);
 exports.readQueryBody = readBody(readQueryPayload);
+exports.getErrorField = getErrorField;
