+++ lib/services/ngsiService.js
@@ -29,6 +29,7 @@ var request = require('request'),
     errors = require('../errors'),
     logger = require('fiware-node-logger'),
     security = require('./securityService'),
+    ngsiParser = require('./ngsiParser'),
     _ = require('underscore'),
     context = {
         op: 'IoTAgentNGSI.NGSIService'
@@ -36,200 +37,6 @@ var request = require('request'),
     registry,
     groupRegistry,
     config;
-
-
-/**
- * Given a NGSI Body, determines whether it contains any NGSI error.
- *
- * @param {String} body             String representing a NGSI body in JSON format.
- * @return {Number|*}
- */
-function getErrorField(body) {
-    var errorField = body.errorCode ||
-        body.orionError;
-
-    if (body && body.contextResponses &&
-        body.contextResponses[0] && body.contextResponses[0] &&
-        body.contextResponses[0].statusCode && body.contextResponses[0].statusCode.code !== '200') {
-        errorField = body.contextResponses[0].statusCode.reasonPhrase +
-        ': ' + body.contextResponses[0].statusCode.details;
-    }
-
-    return errorField;
-}
-
-/**
- * Generates a handler for the registration requests that checks all the possible errors derived from the registration.
- * The parameter information is needed in order to fulfill error information.
- *
- * @param {Boolen} unregister       Indicates whether this registration is an unregistration or register.
- * @param {Object} deviceData       Object containing all the deviceData needed to send the registration.
- * @return {Function}              The generated handler.
- */
-function createRegistrationHandler(unregister, deviceData, callback) {
-    return function handleRegistrationResponse(error, response, body) {
-        if (error) {
-            logger.error(context, 'Connection error sending registrations to the Context Broker: %s', error);
-            callback(error);
-        } else if (response && body && response.statusCode === 200) {
-            var errorField = getErrorField(body);
-
-            if (errorField) {
-                logger.error(context, 'Registration error connecting to the Context Broker: %j', errorField);
-                callback(new errors.BadRequest(JSON.stringify(errorField)));
-            } else {
-                logger.debug(context, 'Registration success.');
-                callback(null, body);
-            }
-        } else {
-            var errorObj;
-
-            logger.error(context, 'Protocol error connecting to the Context Broker: %j', errorObj);
-
-            if (unregister) {
-                errorObj = new errors.UnregistrationError(deviceData.id, deviceData.type);
-            } else {
-                errorObj = new errors.RegistrationError(deviceData.id, deviceData.type);
-            }
-
-            callback(errorObj);
-        }
-    };
-}
-
-/**
- * Sends a Context Provider registration or unregistration request to the Context Broker. As registrations cannot be
- * removed, an unregistration consists of an update of the existent registration to make reduce its duration to
- * 1 second.
- *
- * The entity type, entity name and lazy attributes list are needed in order to send the registration:
- *
- * @param {Boolen} unregister       Indicates whether this registration is an unregistration or register.
- * @param {Object} deviceData       Object containing all the deviceData needed to send the registration.
- */
-function sendRegistrations(unregister, deviceData, callback) {
-    var options = {
-        url: 'http://' + config.contextBroker.host + ':' + config.contextBroker.port + '/NGSI9/registerContext',
-        method: 'POST',
-        json: {
-            contextRegistrations: [
-                {
-                    entities: [
-                        {
-                            type: deviceData.type,
-                            isPattern: 'false',
-                            id: String(deviceData.name)
-                        }
-                    ],
-                    attributes: [],
-                    providingApplication: config.providerUrl
-                }
-            ],
-            duration: (unregister) ? 'PT1S' : config.deviceRegistrationDuration
-        },
-        headers: {
-            'fiware-service': deviceData.service,
-            'fiware-servicepath': deviceData.subservice
-        }
-    };
-
-    if (deviceData.registrationId) {
-        options.json.registrationId = deviceData.registrationId;
-    }
-
-    function formatAttributes(originalVector) {
-        var attributeList = [];
-
-        if (originalVector && originalVector.length) {
-            for (var i = 0; i < originalVector.length; i++) {
-                attributeList.push({
-                    name: originalVector[i].name,
-                    type: originalVector[i].type,
-                    isDomain: 'false'
-                });
-            }
-        }
-
-        return attributeList;
-    }
-
-    options.json.contextRegistrations[0].attributes = options.json.contextRegistrations[0].attributes.concat(
-        formatAttributes(deviceData.lazy),
-        formatAttributes(deviceData.commands)
-    );
-
-    logger.debug(context, 'Sending device registrations to Context Broker at [%s]', options.url);
-    logger.debug(context, 'Using the following request:\n\n%s\n\n', JSON.stringify(options, null, 4));
-
-    request(options, createRegistrationHandler(unregister, deviceData, callback));
-}
-
-/**
- * Process the response from a Register Context request for a device, extracting the 'registrationId' and creating the
- * device object that will be stored in the registry.
- *
- * @param {Object} deviceData       Object containing all the deviceData needed to send the registration.
- *
- */
-function processContextRegistration(deviceData, body, callback) {
-    var newDevice = _.clone(deviceData);
-
-    newDevice.registrationId = body.registrationId;
-
-    callback(null, newDevice);
-}
-
-/**
- * Register a new device identified by the Id and Type in the Context Broker, and the internal registry.
- *
- * The device id and type are required fields for any registration. The rest of the parameters are optional, but, if
- * they are not present in the function call arguments, the type must be registered in the configuration, so the
- * service can infer their default values from the configured type. If an optional attribute is not given in the
- * parameter list and there isn't a default configuration for the given type, a TypeNotFound error is raised.
- *
- * When an optional parameter is not included in the call, a null value must be given in its place.
- *
- * @param {Object} deviceObj                    Object with all the device information (mandatory).
- */
-function registerDevice(deviceObj, callback) {
-    var deviceData = _.clone(deviceObj),
-        fields = ['service', 'subservice', 'lazy', 'internalAttributes', 'active', 'staticAttributes', 'commands'],
-        defaults = [null, null, [], [], [], []];
-
-    logger.debug(context, 'Registering device into NGSI Service:\n%s', JSON.stringify(deviceData, null, 4));
-
-    if (!deviceData.type) {
-        deviceData.type = config.defaultType;
-    }
-
-    if (!deviceData.name) {
-        deviceData.name = deviceData.id + ':' + deviceData.type;
-        logger.debug(context, 'Device name not found, falling back to deviceId:type [%s]', deviceData.name);
-    }
-
-    for (var i = 0; i < fields.length; i++) {
-        if (config.types[deviceData.type] && config.types[deviceObj.type][fields[i]]) {
-            deviceData[fields[i]] =
-                (deviceData[fields[i]]) ? deviceData[fields[i]] : config.types[deviceObj.type][fields[i]];
-        } else {
-            deviceData[fields[i]] = (deviceData[fields[i]]) ? deviceData[fields[i]] : defaults[i];
-        }
-    }
-
-    if (!deviceData.service && config.service) {
-        deviceData.service = config.service;
-    }
-
-    if (!deviceData.subservice && config.subservice) {
-        deviceData.subservice = config.subservice;
-    }
-
-    async.waterfall([
-        async.apply(sendRegistrations, false, deviceData),
-        async.apply(processContextRegistration, deviceData),
-        registry.store
-    ], callback);
-}
 
 /**
  * Generate an operation handler for NGSI-based operations (query and update). The handler takes care of identifiying
@@ -260,7 +67,7 @@ function generateNGSIOperationHandler(operationName, entityName, typeInformation
 
             callback(new errors.BadRequest(body.orionError.details));
         } else if (response && body && response.statusCode === 200) {
-            var errorField = getErrorField(body);
+            var errorField = ngsiParser.getErrorField(body);
 
             if (errorField) {
                 logger.error(
@@ -522,105 +329,6 @@ function setCommandResult(entityName, resource, apikey, commandName,
 }
 
 /**
- * Unregister a device from the Context broker and the internal registry.
- *
- * @param {String} id           Device ID of the device to register.
- */
-function unregisterDevice(id, callback) {
-    function processContextUnregister(body, innerCallback) {
-        innerCallback(null);
-    }
-
-    logger.debug(context, 'Updating device register in NGSI Service');
-
-    async.waterfall([
-        async.apply(registry.get, id),
-        async.apply(sendRegistrations, true),
-        processContextUnregister,
-        async.apply(registry.remove, id)
-    ], function(error) {
-        callback(error);
-    });
-}
-
-/**
- * Updates the register of an existing device identified by the Id and Type in the Context Broker, and the internal
- * registry.
- *
- * The device id and type are required fields for a registration updated. Only the following attributes will be
- * updated: lazy, active and internalId. Any other change will be ignored. The registration for the active attributes
- * of the updated entity will be updated if existing, and created if not.
- *
- * @param {Object} deviceObj                    Object with all the device information (mandatory).
- */
-function updateRegisterDevice(deviceObj, callback) {
-    if (!deviceObj.id || !deviceObj.type) {
-        callback(new errors.MissingAttributes('Id or device missing'));
-        return;
-    }
-
-    logger.debug(context, 'Update provisioned device in NGSI Service');
-
-    function combineWithNewDevice(newDevice, oldDevice, callback) {
-        if (oldDevice) {
-            oldDevice.internalId = newDevice.internalId;
-            oldDevice.lazy = newDevice.lazy;
-            oldDevice.active = newDevice.active;
-            oldDevice.name = newDevice.name;
-            oldDevice.type = newDevice.type;
-            oldDevice.timezone = newDevice.timezone;
-
-            callback(null, oldDevice);
-        } else {
-            callback(new errors.DeviceNotFound(newDevice.id));
-        }
-    }
-
-    async.waterfall([
-        async.apply(registry.get, deviceObj.id),
-        async.apply(combineWithNewDevice, deviceObj),
-        async.apply(sendRegistrations, false),
-        async.apply(processContextRegistration, deviceObj),
-        registry.update
-    ], callback);
-}
-
-/**
- * Return a list of all the devices registered in the system. This function can be invoked in two different ways:
- * with just one parameter (the callback) or with three parameters (including limit and offset).
- *
- * @param {String} service      Service for which the devices are requested.
- * @param {String} subservice   Subservice inside the service for which the devices are requested.
- * @param {Number} limit        Maximum number of entries to return.
- * @param {Number} offset       Number of entries to skip for pagination.
- */
-function listDevices(service, subservice, limit, offset, callback) {
-    if (!callback) {
-        callback = limit;
-    }
-
-    registry.list(service, subservice, limit, offset, callback);
-}
-
-/**
- * Retrieve a device from the device registry.
- *
- * @param {String} deviceId         ID of the device to be found.
- */
-function getDevice(deviceId, callback) {
-    registry.get(deviceId, callback);
-}
-
-/**
- * Retrieve a device from the registry based on its entity name.
- *
- * @param {String} deviceName       Name of the entity associated to a device.
- */
-function getDeviceByName(deviceName, callback) {
-    registry.getByName(deviceName, callback);
-}
-
-/**
  * Initializes the NGSI service. The initialization requires a configuration object and a reference to a device
  * registry.
  *
@@ -636,38 +344,8 @@ function init(newRegistry, newGroupRegistry, newConfig, callback) {
     callback(null);
 }
 
-/**
- * Wraps a function, throwing an exception if the function is invoked before the registry is initialized.
- *
- * @param {Function} fn                 Original function to wrap.
- * @return {Function}                   Wrapped function.
- */
-function checkRegistry(fn) {
-    return function() {
-        var args = Array.prototype.slice.call(arguments),
-            callbacks = args.slice(-1);
-
-        if (registry) {
-            fn.apply(null, args);
-        } else if (callbacks && callbacks.length === 1 && (typeof callbacks[0] === 'function')) {
-            logger.error(context, 'Tried to access device information before a registry was available');
-            callbacks[0](new errors.RegistryNotAvailable());
-        } else {
-            logger.error(context, 'Tried to access device information without providing a callback');
-        }
-    };
-}
-
 exports.init = init;
-exports.register = registerDevice;
-exports.updateRegister = updateRegisterDevice;
-exports.unregister = unregisterDevice;
+
 exports.update = executeWithDeviceInformation(sendUpdateValue);
 exports.query = executeWithDeviceInformation(sendQueryValue);
 exports.setCommandResult = setCommandResult;
-
-exports.listDevices = checkRegistry(listDevices);
-exports.getDevice = checkRegistry(getDevice);
-exports.getDeviceByName = checkRegistry(getDeviceByName);
-
-
